api/requirements.txt
Django==2.0
django-filter==1.1.0
djangorestframework==3.8.2
mysqlclient
pytz==2018.4
mysql-connector-python==8.0.6
django-cors-headers
